Refactor LiveViewScene
