--- factory-ai-vision/EdgeSolution/modules/WebModule/ui/src/components/LiveViewContainer/index.tsx
@@ -1,3 +1,0 @@
-import { LiveViewContainer } from './LiveViewContainer';
-
-export { LiveViewContainer };
