+++ factory-ai-vision/EdgeSolution/modules/WebModule/ui/src/components/LiveViewScene/index.tsx
@@ -1,0 +1,3 @@
+import { LiveViewScene } from './LiveViewScene';
+
+export { LiveViewScene };
